+++ CHANGELOG.md
@@ -80,6 +80,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - target sdk is outdated
 - OpenedListPage : Screen don't blink anymore on item creation/deletion
 - Analytics tags
+- New_profile : user avatar is oval
 
 ### Security